[refactor] rename get-free-slots to get-calendar-free-slots
